// src/neynarV2API/swagger/models/cast-param-type.ts
/* tslint:disable */
/**
 * v2 Farcaster
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * The query param accepted by the API. Sent along with identifier param. url - Cast identifier is a url hash - Cast identifier is a hash
 * @export
 * @enum {string}
 */

export const CastParamType = {
  Url: "url",
  Hash: "hash",
} as const;

// eslint-disable-next-line @typescript-eslint/no-redeclare
export type CastParamType = (typeof CastParamType)[keyof typeof CastParamType];